requirements.txt
streamlit==1.51.0
pandas==2.3.3
PyPDF2==3.0.1
python-docx==1.2.0
langchain==0.2.12
langchain-groq==1.1.0
langchain-huggingface==0.4.0
sentence-transformers==5.1.2
faiss-cpu==1.7.4
